0914将数据插入到original_custom_tag中

+++ marketing-tool-job/src/main/java/cn/rongcapital/mkt/job/service/impl/DataCustomerTagToDataPartyImpl.java
@@ -26,7 +26,7 @@ public class DataCustomerTagToDataPartyImpl extends AbstractDataPartySyncService
 	@Override
 	public int queryTotalCount() {
 		DataCustomerTags dataCustomerTags = new DataCustomerTags();
-		dataCustomerTags.setStatus(StatusEnum.ACTIVE.getStatusCode());
+		dataCustomerTags.setStatus(StatusEnum.ACTIVE.getStatusCode().byteValue());
 		return dataCustomerTagsDao.selectListCount(dataCustomerTags);
 	}
 
@@ -34,7 +34,7 @@ public class DataCustomerTagToDataPartyImpl extends AbstractDataPartySyncService
 	public DataPartySyncVO<Integer> querySyncData(Integer startIndex, Integer pageSize) {
 
 		DataCustomerTags dataCustomerTags = new DataCustomerTags();
-		dataCustomerTags.setStatus(StatusEnum.ACTIVE.getStatusCode());
+		dataCustomerTags.setStatus(StatusEnum.ACTIVE.getStatusCode().byteValue());
 		dataCustomerTags.setPageSize(pageSize);
 		dataCustomerTags.setStartIndex(startIndex);
 		List<DataCustomerTags> dataCustomerTagsList = dataCustomerTagsDao.selectList(dataCustomerTags);
