+++ marketing-tool-job/src/main/java/cn/rongcapital/mkt/job/service/impl/OriginalDataCustomTagScheduleServiceImpl.java
@@ -34,7 +34,7 @@ public class OriginalDataCustomTagScheduleServiceImpl implements OriginalDataCus
     @Transactional(propagation = Propagation.REQUIRED, readOnly = false)
     public void cleanData() {
         OriginalDataCustomerTags paramOriginalDataCustomTags = new OriginalDataCustomerTags();
-        paramOriginalDataCustomTags.setStatus(StatusEnum.ACTIVE.getStatusCode());
+        paramOriginalDataCustomTags.setStatus(StatusEnum.ACTIVE.getStatusCode().byteValue());
         int totalCount = originalDataCustomerTagsDao.selectListCount(paramOriginalDataCustomTags);
         int totalPages = (totalCount + BATCH_NUM - 1) / BATCH_NUM;
         paramOriginalDataCustomTags.setPageSize(BATCH_NUM);
@@ -65,7 +65,7 @@ public class OriginalDataCustomTagScheduleServiceImpl implements OriginalDataCus
             BeanUtils.copyProperties(tmpOriginalDataCustomerTag, paramDataCustomTags);
 
             // 因为在一个事务里 , 直接修改OriginalDataCustomTags的状态
-            tmpOriginalDataCustomerTag.setStatus(StatusEnum.PROCESSED.getStatusCode());
+            tmpOriginalDataCustomerTag.setStatus(StatusEnum.PROCESSED.getStatusCode().byteValue());
             originalDataCustomerTagsDao.updateById(tmpOriginalDataCustomerTag);
             dataCustomerTags.add(paramDataCustomTags);
         }
